rendered all courses from store succesfully
