+++ src/store/store.js
@@ -2,7 +2,7 @@ import { configureStore } from '@reduxjs/toolkit';
 
 import authorsReducer from './authors/authorsSlice';
 import coursesReducer from './courses/coursesSlice';
-import userReducer from './user/userSlice';
+// import userReducer from './user/userSlice';
 
 export const store = configureStore({
 	reducer: {
